avro: Move to own pkg
